+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- PSR-7 request object is available in PSR-14 event (#2)
+
 ## [1.2.0] - 2023-08-07
 
 ### Added